Made it so the upgrade table requires tokens to actually upgrade. The criteria to create a recipe is as follows:
The type required is "rogue:upgrade_table"

The ingredients are in order of item being upgraded, and the item required to upgrade it.

The cost is how many of the second item are required to upgrade.

The output is the upgraded item, and the count for that is how many of that item will be produced.

Potential changes I might make:
Change it to be more like crafing table than a furnace,
thought adding a "failed to upgrade" chance that would only happen after it finished upgrading